feat: refactor `harper-cli` ariadne output adding suggestions
